feat[api-server]: improved friends

+++ modules/api-server/src/api/handlers/friend/delete-friend.ts
@@ -1,31 +1,32 @@
-import { defineHandler, defineSchema } from "api/lib";
-import { Type as type } from "@sinclair/typebox";
+import { defineHandler, getResponse } from "api/lib";
 import { Server } from "platform/types";
 import { friendService } from "services/friend";
 import { inviteService } from "services/invite";
 
-const schema = defineSchema({
-  body: type.Object({
-    userId: type.Number(),
-    friendId: type.Number(),
-  })
-});
 
-async function handler({ server, session, params, body }: Server.Request<typeof schema>, rep: Server.Reply): Promise<Server.Reply> {
-  const firstSideFriendfriend = await friendService.findFriendByUsers(body.userId, body.friendId);
+async function handler({ server, session, params, body }: Server.Request, rep: Server.Reply): Promise<Server.Reply> {
+  const user = session.get("user") as { id: number; } | null;
+  const friendId = (params as { id: number }).id;
+
+  if (!user){
+    throw Error("User is unauthorized");
+  }
+
+  const firstSideFriendfriend = await friendService.getFriend(friendId);
+
   if (!firstSideFriendfriend){
     throw Error("Friend not found");
   }
-  const secondSideFriend = await friendService.findFriendByUsers(body.friendId, body.userId);
 
-  await inviteService.deleteInvite(firstSideFriendfriend.invite.id)
-  await friendService.deleteFriend(firstSideFriendfriend.id)
-  await friendService.deleteFriend(secondSideFriend.id)
+  const secondSideFriend = await friendService.findFriendByUsers(firstSideFriendfriend.friend.id, user.id);
 
-  return rep.status(200).send({ success: true, message: `Friend deleted successfully.` });
+  await friendService.deleteFriend(firstSideFriendfriend.id);
+  await friendService.deleteFriend(secondSideFriend.id);
+  await inviteService.deleteInvite(secondSideFriend.invite.id);
+
+  return rep.status(200).send(getResponse("success", [], `Friend deleted successfully.` ));
 }
 
 export const deleteFriend = defineHandler({
-  schema,
   handler
 });
